+++ gm/pdf_never_embed.cpp
@@ -26,7 +26,7 @@ static void excercise_draw_pos_text(SkCanvas* canvas,
                                     SkScalar x, SkScalar y,
                                     const SkFont& font,
                                     const SkPaint& paint) {
-    const int count = font.countText(text, strlen(text), SkTextEncoding::kUTF8);
+    const size_t count = font.countText(text, strlen(text), SkTextEncoding::kUTF8);
     SkTextBlobBuilder builder;
     auto rec = builder.allocRunPos(font, count);
     font.textToGlyphs(text, strlen(text), SkTextEncoding::kUTF8, {rec.glyphs, count});
